+++ Assets/Standard Resources/ALINE/AlineHDRPCustomPass.cs
@@ -13,13 +13,13 @@ namespace Drawing {
 #if MODULE_RENDER_PIPELINES_HIGH_DEFINITION_9_0_OR_NEWER
 		protected override void Execute (CustomPassContext context) {
 			UnityEngine.Profiling.Profiler.BeginSample("ALINE");
-			DrawingManager.instance.SubmitFrame(context.hdCamera.camera, context.cmd, true);
+			DrawingManager.instance.SubmitFrame(context.hdCamera.camera, new DrawingData.CommandBufferWrapper { cmd = context.cmd }, true);
 			UnityEngine.Profiling.Profiler.EndSample();
 		}
 #else
 		protected override void Execute (ScriptableRenderContext context, CommandBuffer cmd, HDCamera camera, CullingResults cullingResult) {
 			UnityEngine.Profiling.Profiler.BeginSample("ALINE");
-			DrawingManager.instance.SubmitFrame(camera.camera, cmd, true);
+			DrawingManager.instance.SubmitFrame(camera.camera, new DrawingData.CommandBufferWrapper { cmd = cmd }, true);
 			UnityEngine.Profiling.Profiler.EndSample();
 		}
 #endif