+++ Assets/Standard Resources/ALINE/Editor/DrawingSettingsEditor.cs
@@ -19,6 +19,7 @@ namespace Drawing {
 				guiHandler = (searchContext) =>
 				{
 					var settings = new SerializedObject(DrawingSettings.GetSettingsAsset());
+					EditorGUILayout.HelpBox("Opacity of lines, solid objects and text drawn using ALINE. When drawing behind other objects, an additional opacity multiplier is applied.", MessageType.None);
 					EditorGUILayout.Separator();
 					EditorGUILayout.LabelField("Lines", EditorStyles.boldLabel);
 					EditorGUILayout.Slider(settings.FindProperty("settings.lineOpacity"), 0, 1, new GUIContent("Opacity", "Opacity of lines when in front of objects"));
@@ -32,7 +33,8 @@ namespace Drawing {
 					EditorGUILayout.Slider(settings.FindProperty("settings.textOpacity"), 0, 1, new GUIContent("Opacity", "Opacity of text when in front of other objects"));
 					EditorGUILayout.Slider(settings.FindProperty("settings.textOpacityBehindObjects"), 0, 1, new GUIContent("Opacity (occluded)", "Additional opacity multiplier of text when behind or inside other objects"));
 					EditorGUILayout.Separator();
-					EditorGUILayout.HelpBox("Opacity of lines, solid objects and text drawn using ALINE. When drawing behind other objects, an additional opacity multiplier is applied.", MessageType.None);
+					EditorGUILayout.Slider(settings.FindProperty("settings.curveResolution"), 0.1f, 3f, new GUIContent("Curve resolution", "Higher values will make curves smoother, but also a bit slower to draw."));
+
 					settings.ApplyModifiedProperties();
 					if (GUILayout.Button("Reset to default")) {
 						var def = DrawingSettings.DefaultSettings;
@@ -43,6 +45,7 @@ namespace Drawing {
 						current.settings.solidOpacityBehindObjects = def.solidOpacityBehindObjects;
 						current.settings.textOpacity = def.textOpacity;
 						current.settings.textOpacityBehindObjects = def.textOpacityBehindObjects;
+						current.settings.curveResolution = def.curveResolution;
 						EditorUtility.SetDirty(current);
 					}
 				},